Use proper naming and validation on date section in edit project
